+++ tests/unit/includes/gateway/abstract-omise-payment-base-card-test.php
@@ -87,7 +87,7 @@ class Omise_Payment_Base_Card_Test extends Omise_Test_Case {
 		$_POST['omise_token'] = 'tokn_123';
 		$_POST['omise_save_customer_card'] = '';
 
-		$klass = $this->new_instance( [ 'is_passkey_enabled' => 'no' ] );
+		$klass = $this->new_instance();
 		$klass->payment_action = 'auto_capture';
 		$result = $klass->charge( $order_mock->get_id(), $order_mock );
 
@@ -104,58 +104,6 @@ class Omise_Payment_Base_Card_Test extends Omise_Test_Case {
 		);
 		$this->assertEquals( 'charge', $result['object'] );
 		$this->assertEquals( 'chrg_test_no1t4tnemucod0e51mo', $result['id'] );
-	}
-
-	public function test_base_card_charge_with_wc_block_and_passkey_enabled() {
-		$charge_mock = Mockery::mock( 'overload:OmiseCharge' );
-		$charge_mock->shouldReceive( 'create' )->once()->andReturn( [ 'object' => 'charge' ] );
-		$order_mock = $this->get_order_mock( 2000, 'thb' );
-
-		$_POST['omise_token'] = 'tokn_567';
-		$_POST['omise_save_customer_card'] = '';
-		$_POST['wc_block_payment'] = '1';
-
-		$klass = $this->new_instance( [ 'is_passkey_enabled' => 'yes' ] );
-		$klass->payment_action = 'auto_capture';
-		$klass->charge( $order_mock->get_id(), $order_mock );
-
-		$charge_mock->shouldHaveReceived( 'create' )->once()->with(
-			[
-				'amount' => 200000,
-				'currency' => 'THB',
-				'description' => 'WooCommerce Order id 123',
-				'return_uri' => 'https://abc.com/order/complete',
-				'metadata' => [ 'order_id' => 123 ],
-				'card' => 'tokn_567',
-				'capture' => true,
-				'authentication' => 'PASSKEY',
-			]
-		);
-	}
-
-	public function test_base_card_charge_with_wc_shortcode_and_passkey_enabled() {
-		$charge_mock = Mockery::mock( 'overload:OmiseCharge' );
-		$charge_mock->shouldReceive( 'create' )->once()->andReturn( [ 'object' => 'charge' ] );
-		$order_mock = $this->get_order_mock( 100.50, 'thb' );
-
-		$_POST['omise_token'] = 'tokn_567';
-		$_POST['omise_save_customer_card'] = '';
-
-		$klass = $this->new_instance( [ 'is_passkey_enabled' => 'yes' ] );
-		$klass->payment_action = 'auto_capture';
-		$klass->charge( $order_mock->get_id(), $order_mock );
-
-		$charge_mock->shouldHaveReceived( 'create' )->once()->with(
-			[
-				'amount' => 10050,
-				'currency' => 'THB',
-				'description' => 'WooCommerce Order id 123',
-				'return_uri' => 'https://abc.com/order/complete',
-				'metadata' => [ 'order_id' => 123 ],
-				'card' => 'tokn_567',
-				'capture' => true,
-			]
-		);
 	}
 
 	public function test_base_card_result_with_paid_charge() {