+++ tests/unit/includes/gateway/class-omise-payment-internetbanking-test.php
@@ -4,6 +4,9 @@ use Brain\Monkey;
 
 require_once __DIR__ . '/class-omise-offsite-test.php';
 
+/**
+ * @runTestsInSeparateProcesses
+ */
 class Omise_Payment_Internetbanking_Test extends Omise_Offsite_Test
 {
     protected function setUp(): void
@@ -12,6 +15,7 @@ class Omise_Payment_Internetbanking_Test extends Omise_Offsite_Test
         parent::setUp();
         Monkey\Functions\expect('wp_kses');
         Monkey\Functions\expect('add_action');
+        Monkey\Functions\stubs(['sanitize_text_field' => 'Sanitized text']);
         require_once __DIR__ . '/../../../../includes/gateway/class-omise-payment-internetbanking.php';
     }